Little kuldy, but fits all three example case
