+++ tests/test_c.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function
 
-import c
+import tools.c as c
 
 def test_c():
     assert c.do_something() == "|config|"
