+++ tests/test_d.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function
 
-import d
+import tools.d as d
 
 
 def test_d():
