+++ tests/test_f.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function
 
-import f
+import tools.f as f
 
 
 def test_f():
