+++ tests/test_k.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function
 
-import k
+import tools.k as k
 
 
 def test_k():
