+++ tests/test_my_tool.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function
 
-import my_tool
+import tools.my_tool as my_tool
 
 
 def test_my_tool():
