+++ tests/test_s.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function
 
-import s
+import tools.s as s
 
 
 def test_my_tool():
