+++ tests/test_v.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function
 
-import v
+import tools.v as v
 
 
 def test_v():
