+++ tools/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+# -*- coding: UTF-8 -*-
+
