+++ tools/d.py
