+++ tools/s.py
@@ -1,16 +1,18 @@
 #!/usr/bin/env python
 # coding=utf-8
 from __future__ import absolute_import, division, print_function
-import v
-def do_something():
-    data = ""
-    with open("resources/lol.txt") as f:
-        data = f.read().strip()
-    return "|keys|" + data + v.do_something()
 
-print("K!")
+try:
+    import t
+except ImportError:
+    import tools.t as t
+
+def do_something():
+    return t.do_something() + "|installed|"
+
+print("S!")
 
 if __name__ == "__main__":
-    print("K main!")
+    print("S main!")
     print(do_something())
