+++ tools/t.py
